update folder/file names
