+++ client/src/components/MainRoute.js
@@ -1,13 +1,13 @@
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import Movies from './Movies'
+import Videos from './Videos'
 import PlayVideo from './PlayVideo'
 
 function MainRoute() {
     return (
         <Router>
             <Switch>
-                <Route exact path="/" component={Movies} />
-                <Route path="/video" component={PlayVideo}/>
+                <Route exact path="/" component={Videos} />
+                {/* <Route path="/video" component={PlayVideo}/> */}
             </Switch>
         </Router>
     )
